+++ EasyDriver/changelog.md
@@ -5,6 +5,8 @@
 - [ ] Auto generating markdown docs
 - [ ] Logger compatible with Xunit
 - [ ] Auto download browser using selenium-manager
+- [x] Get Texts using Js experimental option
+- [x] Add method: MapUsingJs
 - [x] Rename `S() -> Locator()`, `_S() -> SubLocator()` keeping aliases: `S`, `_S`
 - [x] Fix IsDisplayed throwing StaleElementReference exception